chore
cleaning up some of the NetworkVariable associated tests so they are contained in a subfolder and so each test is contained in its own CS file.
