+++ com.unity.netcode.gameobjects/Tests/Runtime/NetworkVariable.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7b4da27c6efa9684893f85f5d3ad80e6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
