+++ com.unity.netcode.gameobjects/Tests/Runtime/NetworkVariable/NetworkVariableInheritanceTests.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 41d4aef8f33a8eb4e87879075f868e66
