+++ com.unity.netcode.gameobjects/Tests/Runtime/NetworkVariable/NetworkVariablePermissionTests.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 60d49d322bef8ff4ebb8c4abf57e18e3
